Moved and renamed MessageHistoryTextBox.cs for better fit.
