+++ osu.Game/Overlays/Chat/ChatTextBox.cs
@@ -2,11 +2,12 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using osu.Framework.Bindables;
+using osu.Game.Graphics.UserInterface;
 using osu.Game.Resources.Localisation.Web;
 
 namespace osu.Game.Overlays.Chat
 {
-    public class ChatTextBox : ChatRecentTextBox
+    public class ChatTextBox : MessageHistoryTextBox
     {
         public readonly BindableBool ShowSearch = new BindableBool();
